refactor(v1): make Dialect the single place a protocol lives
Move every protocol-specific operation onto the `Dialect` so adding a harness with a new
wire format is one self-contained module + a `DIALECTS` entry — the proxy and interception
server are now fully generic over the interface:

- `upstream_path` + `apply_overrides(body, model, sampling)`: the proxy forwards the request
  byte-exact and the dialect imposes only what the eval owns — model overlays, sampling is
  authoritative (the program's sampling keys are dropped, the eval's applied), in the
  protocol's own shape. Fixes the proxy hardcoding `/chat/completions` and the chat-shaped
  sampling overlay.
- `serialize_response` + `extend`: the dialect owns the two vf -> wire cases (the renderer's
  generated response; user-sim turn injection), so `interception/server.py` has zero
  wire-format code — it drops the chat-shaped `raw_messages`/`["choices"][0]` handling.
- Wire serializers (`message_to_wire`/`tool_to_wire`) live in the chat dialect (the chat-only
  renderer + default harness import them from there); `model_error` moves to `errors` (its
  natural home, shared by both clients). `clients/proxy.py` is now just `ProxyClient`.

Keep `AsyncOpenAI` as the proxy transport: it's used as a raw poster and handles the
OpenAI-compatible endpoint family (base url, Bearer auth, billing headers) every OpenAI-SDK
dialect shares; a non-OpenAI provider is a separate transport axis, not a dialect concern.
